Event weight int to double fix. Also removed hd3 related arrays in SigMc Readers.

+++ create_database/McEventSQLiteTableBuilder.h
@@ -14,7 +14,7 @@ class McEventSQLiteTableBuilder : public EventSQLiteTableBuilder {
   private:
     int mc_evttype;
     int sp_mode;
-    int event_weight;
+    double event_weight;
 
     std::map<std::string, std::pair<std::string, std::string> > sample_assignment_map;
     void BuildSampleAssignmentMap(const char *assignment_fname);
@@ -28,7 +28,7 @@ class McEventSQLiteTableBuilder : public EventSQLiteTableBuilder {
 
     void set_mc_evttype(int evttype) { mc_evttype = evttype; }
     void set_sp_mode(int spmode) { sp_mode = spmode; }
-    void set_event_weight(int evt_wgt) { event_weight = evt_wgt; }
+    void set_event_weight(double evt_wgt) { event_weight = evt_wgt; }
 };
 
 #endif
